+++ cmd/argo/commands/list.go
@@ -9,8 +9,8 @@ import (
 	"text/tabwriter"
 	"time"
 
+	"github.com/argoproj/pkg/humanize"
 	argotime "github.com/argoproj/pkg/time"
-	humanize "github.com/dustin/go-humanize"
 	"github.com/spf13/cobra"
 	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -102,18 +102,6 @@ func NewListCommand() *cobra.Command {
 	return command
 }
 
-var timeMagnitudes = []humanize.RelTimeMagnitude{
-	{D: time.Second, Format: "0s", DivBy: time.Second},
-	{D: 2 * time.Second, Format: "1s %s", DivBy: 1},
-	{D: time.Minute, Format: "%ds %s", DivBy: time.Second},
-	{D: 2 * time.Minute, Format: "1m %s", DivBy: 1},
-	{D: time.Hour, Format: "%dm %s", DivBy: time.Minute},
-	{D: 2 * time.Hour, Format: "1h %s", DivBy: 1},
-	{D: humanize.Day, Format: "%dh %s", DivBy: time.Hour},
-	{D: 2 * humanize.Day, Format: "1d %s", DivBy: 1},
-	{D: humanize.Week, Format: "%dd %s", DivBy: humanize.Day},
-}
-
 func printTable(wfList []wfv1.Workflow, listArgs *listFlags) {
 	w := tabwriter.NewWriter(os.Stdout, 0, 0, 3, ' ', 0)
 	if listArgs.allNamespaces {
@@ -121,20 +109,19 @@ func printTable(wfList []wfv1.Workflow, listArgs *listFlags) {
 	}
 	fmt.Fprint(w, "NAME\tSTATUS\tAGE\tDURATION")
 	if listArgs.output == "wide" {
-		fmt.Fprint(w, "\tR/C\tPARAMETERS")
+		fmt.Fprint(w, "\tP/R/C\tPARAMETERS")
 	}
 	fmt.Fprint(w, "\n")
 	for _, wf := range wfList {
-		cTime := time.Unix(wf.ObjectMeta.CreationTimestamp.Unix(), 0)
-		ageStr := humanize.CustomRelTime(cTime, time.Now(), "", "", timeMagnitudes)
-		durationStr := humanizeDurationShort(wf.Status.StartedAt, wf.Status.FinishedAt)
+		ageStr := humanize.RelativeDurationShort(wf.ObjectMeta.CreationTimestamp.Time, time.Now())
+		durationStr := humanize.RelativeDurationShort(wf.Status.StartedAt.Time, wf.Status.FinishedAt.Time)
 		if listArgs.allNamespaces {
 			fmt.Fprintf(w, "%s\t", wf.ObjectMeta.Namespace)
 		}
 		fmt.Fprintf(w, "%s\t%s\t%s\t%s", wf.ObjectMeta.Name, worklowStatus(&wf), ageStr, durationStr)
 		if listArgs.output == "wide" {
-			running, completed := countCompletedRunning(&wf)
-			fmt.Fprintf(w, "\t%d/%d", running, completed)
+			pending, running, completed := countPendingRunningCompleted(&wf)
+			fmt.Fprintf(w, "\t%d/%d/%d", pending, running, completed)
 			fmt.Fprintf(w, "\t%s", parameterString(wf.Spec.Arguments.Parameters))
 		}
 		fmt.Fprintf(w, "\n")
@@ -142,22 +129,24 @@ func printTable(wfList []wfv1.Workflow, listArgs *listFlags) {
 	_ = w.Flush()
 }
 
-func countCompletedRunning(wf *wfv1.Workflow) (int, int) {
-	completed := 0
+func countPendingRunningCompleted(wf *wfv1.Workflow) (int, int, int) {
+	pending := 0
 	running := 0
+	completed := 0
 	for _, node := range wf.Status.Nodes {
-		if len(node.Children) > 0 {
-			// not a pod
-			// TODO: this will change after DAG implementation
+		tmpl := wf.GetTemplate(node.TemplateName)
+		if tmpl == nil || !tmpl.IsPodType() {
 			continue
 		}
 		if node.Completed() {
 			completed++
-		} else {
+		} else if node.Phase == wfv1.NodeRunning {
 			running++
+		} else {
+			pending++
 		}
 	}
-	return running, completed
+	return pending, running, completed
 }
 
 // parameterString returns a human readable display string of the parameters, truncating if necessary
@@ -213,9 +202,9 @@ func worklowStatus(wf *wfv1.Workflow) wfv1.NodePhase {
 			return "Running (Suspended)"
 		}
 		return wf.Status.Phase
-	case "":
+	case "", wfv1.NodePending:
 		if !wf.ObjectMeta.CreationTimestamp.IsZero() {
-			return "Pending"
+			return wfv1.NodePending
 		}
 		return "Unknown"
 	default: