+++ cmd/argo/commands/submit.go
@@ -2,17 +2,21 @@ package commands
 
 import (
 	"bufio"
-	"encoding/json"
+	"fmt"
 	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	"strings"
 
+	"github.com/argoproj/pkg/json"
+	"github.com/spf13/cobra"
+
 	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	cmdutil "github.com/argoproj/argo/util/cmd"
 	"github.com/argoproj/argo/workflow/common"
-	"github.com/spf13/cobra"
+	"github.com/argoproj/argo/workflow/validate"
+	"github.com/ghodss/yaml"
 )
 
 type submitFlags struct {
@@ -21,8 +25,10 @@ type submitFlags struct {
 	instanceID     string   // --instanceid
 	entrypoint     string   // --entrypoint
 	parameters     []string // --parameter
+	parameterFile  string   // --parameter-file
 	output         string   // --output
 	wait           bool     // --wait
+	strict         bool     // --strict
 	serviceAccount string   // --serviceaccount
 }
 
@@ -45,8 +51,10 @@ func NewSubmitCommand() *cobra.Command {
 	command.Flags().StringVar(&submitArgs.generateName, "generate-name", "", "override metadata.generateName")
 	command.Flags().StringVar(&submitArgs.entrypoint, "entrypoint", "", "override entrypoint")
 	command.Flags().StringArrayVarP(&submitArgs.parameters, "parameter", "p", []string{}, "pass an input parameter")
+	command.Flags().StringVarP(&submitArgs.parameterFile, "parameter-file", "f", "", "pass a file containing all input parameters")
 	command.Flags().StringVarP(&submitArgs.output, "output", "o", "", "Output format. One of: name|json|yaml|wide")
 	command.Flags().BoolVarP(&submitArgs.wait, "wait", "w", false, "wait for the workflow to complete")
+	command.Flags().BoolVar(&submitArgs.strict, "strict", true, "perform strict workflow validation")
 	command.Flags().StringVar(&submitArgs.serviceAccount, "serviceaccount", "", "run all pods in the workflow using specified serviceaccount")
 	command.Flags().StringVar(&submitArgs.instanceID, "instanceid", "", "submit with a specific controller's instance id label")
 	return command
@@ -61,7 +69,7 @@ func SubmitWorkflows(filePaths []string, submitArgs *submitFlags) {
 		if err != nil {
 			log.Fatal(err)
 		}
-		workflows = unmarshalWorkflows(body)
+		workflows = unmarshalWorkflows(body, submitArgs.strict)
 	} else {
 		for _, filePath := range filePaths {
 			var body []byte
@@ -82,7 +90,7 @@ func SubmitWorkflows(filePaths []string, submitArgs *submitFlags) {
 					log.Fatal(err)
 				}
 			}
-			wfs := unmarshalWorkflows(body)
+			wfs := unmarshalWorkflows(body, submitArgs.strict)
 			workflows = append(workflows, wfs...)
 		}
 	}
@@ -102,13 +110,17 @@ func SubmitWorkflows(filePaths []string, submitArgs *submitFlags) {
 }
 
 // unmarshalWorkflows unmarshals the input bytes as either json or yaml
-func unmarshalWorkflows(wfBytes []byte) []wfv1.Workflow {
+func unmarshalWorkflows(wfBytes []byte, strict bool) []wfv1.Workflow {
 	var wf wfv1.Workflow
-	err := json.Unmarshal(wfBytes, &wf)
+	var jsonOpts []json.JSONOpt
+	if strict {
+		jsonOpts = append(jsonOpts, json.DisallowUnknownFields)
+	}
+	err := json.Unmarshal(wfBytes, &wf, jsonOpts...)
 	if err == nil {
 		return []wfv1.Workflow{wf}
 	}
-	yamlWfs, err := splitYAMLFile(wfBytes)
+	yamlWfs, err := common.SplitYAMLFile(wfBytes, strict)
 	if err == nil {
 		return yamlWfs
 	}
@@ -132,7 +144,7 @@ func submitWorkflow(wf *wfv1.Workflow, submitArgs *submitFlags) (string, error) 
 		labels[common.LabelKeyControllerInstanceID] = submitArgs.instanceID
 		wf.SetLabels(labels)
 	}
-	if len(submitArgs.parameters) > 0 {
+	if len(submitArgs.parameters) > 0 || submitArgs.parameterFile != "" {
 		newParams := make([]wfv1.Parameter, 0)
 		passedParams := make(map[string]bool)
 		for _, paramStr := range submitArgs.parameters {
@@ -147,6 +159,49 @@ func submitWorkflow(wf *wfv1.Workflow, submitArgs *submitFlags) (string, error) 
 			newParams = append(newParams, param)
 			passedParams[param.Name] = true
 		}
+
+		// Add parameters from a parameter-file, if one was provided
+		if submitArgs.parameterFile != "" {
+			var body []byte
+			var err error
+			if cmdutil.IsURL(submitArgs.parameterFile) {
+				response, err := http.Get(submitArgs.parameterFile)
+				if err != nil {
+					log.Fatal(err)
+				}
+				body, err = ioutil.ReadAll(response.Body)
+				_ = response.Body.Close()
+				if err != nil {
+					log.Fatal(err)
+				}
+			} else {
+				body, err = ioutil.ReadFile(submitArgs.parameterFile)
+				if err != nil {
+					log.Fatal(err)
+				}
+			}
+
+			yamlParams := map[string]interface{}{}
+			err = yaml.Unmarshal(body, &yamlParams)
+			if err != nil {
+				log.Fatal(err)
+			}
+
+			for k, v := range yamlParams {
+				value := fmt.Sprintf("%v", v)
+				param := wfv1.Parameter{
+					Name:  k,
+					Value: &value,
+				}
+				if _, ok := passedParams[param.Name]; ok {
+					// this parameter was overridden via command line
+					continue
+				}
+				newParams = append(newParams, param)
+				passedParams[param.Name] = true
+			}
+		}
+
 		for _, param := range wf.Spec.Arguments.Parameters {
 			if _, ok := passedParams[param.Name]; ok {
 				// this parameter was overridden via command line
@@ -162,7 +217,7 @@ func submitWorkflow(wf *wfv1.Workflow, submitArgs *submitFlags) (string, error) 
 	if submitArgs.name != "" {
 		wf.ObjectMeta.Name = submitArgs.name
 	}
-	err := common.ValidateWorkflow(wf)
+	err := validate.ValidateWorkflow(wf)
 	if err != nil {
 		return "", err
 	}